Refactoring most important road to Business Logic, Data Logic, Backend Logic
